Fix checkpoint bug

+++ test/test_model_checkpoint.py
@@ -4,6 +4,7 @@ import os
 import sys
 import hydra
 import shutil
+import torch
 
 DIR = os.path.dirname(os.path.realpath(__file__))
 ROOT = os.path.join(DIR, "..")
@@ -30,6 +31,17 @@ def remove(path):
         pass
 
 
+class MockModel(torch.nn.Module):
+    """ Mock mdoel that does literaly nothing but holds a state
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.state = torch.nn.parameter.Parameter(torch.tensor([1.0]))
+        self.optimizer = torch.nn.Module()
+        self.schedulers = {}
+
+
 class TestModelCheckpoint(unittest.TestCase):
     def setUp(self):
         self.data_config = OmegaConf.load(os.path.join(DIR, "test_config/data_config.yaml"))
@@ -37,16 +49,16 @@ class TestModelCheckpoint(unittest.TestCase):
         scheduler_config = OmegaConf.load(os.path.join(DIR, "test_config/scheduler_config.yaml"))
         params = load_config("segmentation", "pointnet2")
         self.config = OmegaConf.merge(training_config, scheduler_config, params)
+        self.model_name = "model"
 
     def test_model_ckpt_using_pointnet2ms(self,):
         # Create a checkpt
-        name = "model"
+
         self.run_path = os.path.join(DIR, "checkpt")
-        print(self.run_path)
         if not os.path.exists(self.run_path):
             os.makedirs(self.run_path)
 
-        model_checkpoint = ModelCheckpoint(self.run_path, name, "test", run_config=self.config, resume=False)
+        model_checkpoint = ModelCheckpoint(self.run_path, self.model_name, "test", run_config=self.config, resume=False)
         dataset = MockDatasetGeometric(5)
         model = instantiate_model(self.config, dataset)
         model.set_input(dataset[0], "cpu")
@@ -56,7 +68,7 @@ class TestModelCheckpoint(unittest.TestCase):
         model_checkpoint.save_best_models_under_current_metrics(model, mock_metrics)
 
         # Load checkpoint and initialize model
-        model_checkpoint = ModelCheckpoint(self.run_path, name, "test", self.config, resume=True)
+        model_checkpoint = ModelCheckpoint(self.run_path, self.model_name, "test", self.config, resume=True)
         model2 = model_checkpoint.create_model(dataset, weight_name="acc")
 
         self.assertEqual(str(model.optimizer.__class__.__name__), str(model2.optimizer.__class__.__name__))
@@ -64,9 +76,36 @@ class TestModelCheckpoint(unittest.TestCase):
         self.assertEqual(model.schedulers["lr_scheduler"].state_dict(), model2.schedulers["lr_scheduler"].state_dict())
         self.assertEqual(model.schedulers["bn_scheduler"].state_dict(), model2.schedulers["bn_scheduler"].state_dict())
 
-        shutil.rmtree(self.run_path)
-        remove(os.path.join(ROOT, "{}.pt".format(name)))
-        remove(os.path.join(DIR, "{}.pt".format(name)))
+        remove(os.path.join(ROOT, "{}.pt".format(self.model_name)))
+        remove(os.path.join(DIR, "{}.pt".format(self.model_name)))
+
+    def test_best_metric(self):
+        self.run_path = os.path.join(DIR, "checkpt")
+        if not os.path.exists(self.run_path):
+            os.makedirs(self.run_path)
+
+        model_checkpoint = ModelCheckpoint(self.run_path, self.model_name, "test", run_config=self.config, resume=False)
+        model = MockModel()
+        optimal_state = model.state.item()
+        mock_metrics = {"current_metrics": {"acc": 12}, "stage": "test", "epoch": 10}
+        model_checkpoint.save_best_models_under_current_metrics(model, mock_metrics)
+        model.state[0] = 2
+        mock_metrics = {"current_metrics": {"acc": 0}, "stage": "test", "epoch": 11}
+        model_checkpoint.save_best_models_under_current_metrics(model, mock_metrics)
+        mock_metrics = {"current_metrics": {"acc": 10}, "stage": "train", "epoch": 11}
+        model_checkpoint.save_best_models_under_current_metrics(model, mock_metrics)
+        mock_metrics = {"current_metrics": {"acc": 15}, "stage": "train", "epoch": 11}
+        model_checkpoint.save_best_models_under_current_metrics(model, mock_metrics)
+
+        ckp = torch.load(os.path.join(self.run_path, self.model_name + ".pt"))
+
+        self.assertEqual(ckp["models"]["best_acc"]["state"].item(), optimal_state)
+        self.assertEqual(ckp["models"]["latest"]["state"].item(), model.state.item())
+
+    def tearDown(self):
+        if os.path.exists(self.run_path):
+            shutil.rmtree(self.run_path)
+            # os.remove(os.path.join(DIR, "{}.pt".format(self.model_name)))
 
 
 if __name__ == "__main__":
